change varchar to nvarchar

+++ VloveImport/VloveImport.data/ScrapingData.cs
@@ -27,7 +27,7 @@ namespace VloveImport.data
         public string CUS_BK_ITEMDESC { get; set; }   
         public int CUS_BK_NUMBER { get; set; }
         public int CUS_BK_AMOUNT { get; set; }
-        public float CUS_BK_PRICE { get; set; }
+        public double CUS_BK_PRICE { get; set; }
         public string CUS_BK_SIZE { get; set; }
         public string CUS_BK_COLOR { get; set; }   
         public string CUS_BK_REMARK { get; set; }
